Moved all of the modules back into the main file because it was being annoying
